+++ apps/obsidian/src/utils/validateNodeType.ts
